Better error message for user
If the preset could not be converted we print an error message per
javascript that is visible to the user. This can further be improved by
passing up error messages from the rust code to javascript. For now, we
just know that some error happened and we refer to the console.log().